# Persist SDK Endpoint payloads in Mongo