--- a/packages/back-end/src/events/handlers/webhooks/event-webhooks-utils.ts
+++ b/packages/back-end/src/events/handlers/webhooks/event-webhooks-utils.ts
@@ -1,7 +1,7 @@
 import { createHmac } from "crypto";
 import fetch, { RequestInfo, RequestInit, Response } from "node-fetch";
 import { OrganizationInterface } from "../../../../types/organization";
-import { getApiFeatureObj, GroupMap } from "../../../services/features";
+import { getApiFeatureObj } from "../../../services/features";
 import {
   FeatureCreatedNotificationEvent,
   FeatureDeletedNotificationEvent,
@@ -15,6 +15,7 @@ import {
 } from "../../base-types";
 import { ApiFeatureInterface } from "../../../../types/api";
 import { logger } from "../../../util/logger";
+import { GroupMap } from "../../../../types/saved-group";
 
 export type EventWebHookSuccessResult = {
   result: "success";