### Overview

SDK Endpoint payloads are regenerated from scratch on every request to `/api/features/...`.  This is expensive and slow (averaging ~200ms in production).  GrowthBook Cloud has a CDN in front, which helps, but the cache hit ratio is <95%, so there are still a significant number of requests serving directly from the API.  Plus, self-hosted deployments do not have a CDN by default and thus suffer from slow performance unless custom caching is implemented.

This PR persists generated SDK endpoint payloads in Mongo.  Every time a feature is updated, the cache entries are immediately updated to match.  The SDK endpoints (as well as webhooks) will prefer the cached copy in Mongo and fallback to the existing method.

This also sets the stage for future performance enhancements such as S3 caching and a Proxy Server

### Changes
- New SDKPayloads collection in Mongo to store cached feature definition payloads for each project/environment combo
- Add cache layer to SDK endpoints and webhooks to use this new Mongo collection when possible
- Consolidate all `onFeatureUpdate` logic into a single function call and moved it to the FeatureModel methods so we don't forget them in the future.  Also added `onFeatureCreate` and `onFeatureDelete` placeholder methods.
- Add smarter logic to determine exactly which environment/project pairs are affected by a change
- When updating a savedGroup, check which rules are actually referencing it instead of just marking every feature and every environment as updated
- Add unit tests for the SDK payload generation logic and helper methods

### Benchmarks

When hitting an SDK endpoint locally:
- Old: **80ms**
- New: **16ms**

Production:
- Old: **200ms**
- New: ??? (expected: 40ms)

### Manual Testing
- [x] SDK payload cache is populated when hitting the SDK endpoint
- [x] Changing a feature updates the SDK payload cache, fires webhooks, and logs the event in Mongo
  - [x] Creating a feature
  - [x] Toggling a feature in an evironment
  - [x] Deleting a feature
  - [x] Archiving a feature
  - [x] Updating a saved group
  - [x] Adding a rule
  - [x] Editing a rule
  - [x] Re-ordering rules
  - [x] Deleting a rule
  - [x] Publishing
  - [x] Discarding a draft
  - [x] Toggling via the REST api
  - [x] Scheduled rule cronjob
- [x] SDK payload cache not updated while feature is archived
- [x] SDK payload cache not updated for an environment when the feature i s toggled off
- [x] SDK payload cache updated when a feature is in a project